5단계 - 로또(수동) (#578)
* refactor: add LottoPayment.purchasableAmount to the constructor and modify constructor's access modifier

* docs: add additional features

* refactor: add LottoPayment.manualAmount

- add StringToIntegerConverter
- add MoneyValidator
- rename StringToPaymentConverter to MoneyValidator

* feat: add LottoPresenter.requestManualAmount

* feat: add LottoTickets.createManually() and LottoTickets.merge()

* feat: add LottoPresenter.requestManualNumbersList()

* refactor: change Prize.prize integer to long and modify LottoPresenter.printPaymentResult

* fix: add a validation logic to check duplication of BonusBall

* fix: add a validation logic to check duplication of LottoNumbers

+++ src/main/java/lotto/service/PrizeCalculator.java
@@ -6,12 +6,12 @@ import lotto.domain.Prize;
 
 public class PrizeCalculator {
     public static double returnOfRatio(final int payment, final Map<Prize, Integer> prizeMap) {
-        final int income = sumPrizes(prizeMap);
+        final long income = sumPrizes(prizeMap);
         return Double.valueOf(income) / Double.valueOf(payment);
     }
 
-    private static int sumPrizes(final Map<Prize, Integer> prizeMap) {
-        int sum = 0;
+    private static long sumPrizes(final Map<Prize, Integer> prizeMap) {
+        long sum = 0;
         for (Entry<Prize, Integer> entry : prizeMap.entrySet()) {
             sum += entry.getKey().getPrize() * entry.getValue();
         }
